requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.4
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via aiohttp
charset-normalizer==3.0.1
    # via aiohttp
colorama==0.4.6
    # via loguru
discord-py==2.1.1
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
loguru==0.6.0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
win32-setctime==1.1.0
    # via loguru
yarl==1.8.2
    # via aiohttp
